+++ examples/data-sources/pingfederate_keypairs_signing_csr/data-source.tf
@@ -1,3 +1,0 @@
-data "pingfederate_keypairs_signing_csr" "signingCsr" {
-  keypair_id = "mysigningkeypair"
-}