+++ examples/resources/pingfederate_keypairs_signing_csr_export/resource.tf
@@ -1,0 +1,3 @@
+resource "pingfederate_keypairs_signing_csr_export" "signingCsr" {
+  keypair_id = "mysigningkeypair"
+}